Updated openstack/openstack Project: openstack/python-cinderclient  af3b039cd2529eb048c0bf171ecbb0c313f0d4ed
Remove direct dependency on requests

keystoneauth is the library that pulls in requests. If we have client
libraries also depend directly on requests, we can get into situations
for end users in installations where, because of sequencing of client
library releases there are conflicting versions of requests needed and
pkg_resources/entrypoints break.

Remove the direct depend and let keystoneauth pull it in for us.

Change-Id: If82ee8674b2166ec66c2a978827a203786fabec7