Updated openstack/openstack Project: openstack/governance  d24b5bc66c3078328bcd2318c39af1b63dd739e0
Confirm Thierry Carrez as chair

Hi! Despite being away for the first meeting of the new
membership, I'd like to propose myself to continue chairing
the OpenStack Technical Committee in the coming 6 months.

We need to go through a number of significant changes, to
adapt OpenStack to the new challenges created by this
post-hype phase we are currently in. That will likely involve
an increase in the number of janitorial tasks which traditionally
fell on the Technical Committee chair. As part of my job at the
OpenStack Foundation, I'm able to continue dedicating substantial
time tracking efforts and pushing for their completion. I hope
you will trust me to help us all through those needed changes.

Note: Other candidates should propose alternate patches to this
one, and we can use Gerrit votes to select the winner.

Change-Id: I8b6464ca96f5c04e76794aee05edacaaf4404cfd