Updated openstack/openstack Project: openstack/neutron  96d13e42f49a700f5eab24dde22094b36553249f
Verify metering label exists before applying rule

If the metering-agent receives a label rule before it
has added the label, it will fail to update the iptables
rules as there are no existing chains.

When the action is "create", check if there is an existing
label, and create one and the corresponding iptables chains,
before trying to add the rule.

Closes-Bug: #1617248

Change-Id: Ia0ec1361188cca53023667d249c2b1e10bc22089