Updated openstack/openstack Project: openstack-infra/elastic-recheck  941d6c25f5816f818c96d36eaff9f60d5d7dca3a
Add query for UnicodeDecodeError with journal log handler bug 1580728

The cells job has started failing at a pretty high rate since devstack
was changed to use systemd by default on 5/1 and that's using the new
OSJournalHandler in oslo.log 3.26.0.

I don't know why this primarily affects the cells job though.

135 hits since 5/1, check and gate, master branch only. It's not 100%
failure probably because sometimes when this happens it's during
the delete of a server in tempest cleanup and the test no longer cares
about the state of the instance.

Change-Id: Ib07db71f8675fc6b4ea48f8f62dd0a8a7a1880d8
Related-Bug: #1580728