Updated openstack/openstack Project: openstack/neutron  617fc6087b51709cc8f33367e0054eb4b8f82045
Update auto-addresses on MAC change

On MAC update add an entry for auto-addressed IPs so
that they get removed and re-allocated.

Also add a check to ensure the length of fixed_ips in the result
is what was expected i.e. ensure we replaced the IP's(as is the current
behaviour) and not just added to the list.

Change-Id: Id1eadce7a18ba82d6c1a42f1afb0267f076f9308
Closes-Bug: #1671548