Updated openstack/openstack Project: openstack/ironic-specs  9a83b0deee21298cee6495e85501f28fc4df0ea7
Remove Pike priorities that cannot make it due to recent team changes

Also add backup assignees for folks who are unlikely to be able to continue
working on ironic upstream, and remove folks who resigned from the team.

Lower priority of rescue, as it's in the process of changing assignees,
and thus is at risk.

Change-Id: Ia1d4f0fbe0dd645241faa23d636adc7a61bc81eb