Updated openstack/openstack Project: openstack/training-guides  6483df4b516ca62b0a6357ca0ad23db7c35792cf
[upstream] fix missing line and describe voting/non-voting

While reviewing the content today I noticed that it appears
that I cannot have an image and text on a slide.  I created
a new slide and moved the line.  I also got feedback while
running through this deck that we should add a description of
voting and non-voting jobs.  So, that content is also added.

Change-Id: I7dfce3c5600c995114ae54946fd27e3a5258d6a6