Updated openstack/openstack Project: openstack-infra/shade  184cc232c9d3d4c44300eda04664a43d7d76f5ce
Move REST error_messages to error_message argument

The with shade_exceptions is not doing what is expected here - since it
passes OpenStackCloudException objects on unmodified. Instead, use the
new error_message parameter to continue to give friendly error messages.

Also - wow, we were mocking _compute_client for the nova extensions
tests. I believe that was become those pre-dated our use of
requests_mock. Anyway, that's super lame (and also broke because of this
patch), so it's now fixed.

Change-Id: Ia764cc76d724a8a732fa2b1209a138634a25a9e1