Updated openstack/openstack Project: openstack-infra/puppet-ansible  03b90b3a98401a946d14bcfc8f9ced5e00103c00
Fix beaker on xenial

Since the beaker jobs are being run on xenial, we need a special nodeset
for it, otherwise beaker gives an error:

 beaker-hostgenerator was not able to use this value as input.
 Exiting with an Error.

We also want to install puppet from the Ubuntu repos rather than from
puppetlabs, since puppetlabs doesn't support puppet 3 for Xenial. For
centos we can keep the install process the same.

Change-Id: Ifd2244ae9dd212b2475f9cd6adb994bc058a4769
Depends-On: Ife2993b07dd995fc97676c002a91d0e695407522