Updated openstack/openstack Project: openstack-infra/project-config  e1a3b80aeb7610c20d5d9b01e94500c9843f3c13
Add specs and validation jobs for service-types-authority

There is a patch to service-types-authority to add docs publication and
validation checks. Add jobs to support that. We should wait until the
service-types-authority change has at least some amount of approval, but
that change depends-on this one, so the change itself should be self
gating.

Change-Id: I5eb1503f815440cac2d127c1f1fec1318e894323
Needed-by: I0f5023d28d9344e87db5a01e7fd1f19a427bec4c