Updated openstack/openstack Project: openstack/python-openstackclient  07a4363e995716f9b898623d9e43746fd36b5172
Skip floating ip attach functional test on nova-net

As of Ocata release Nova forces nova-network to run in a cells v1
configuration.  Floating IP and network functions currently do not
work in the gate jobs so we have to skip this.  It is known to work
tested against a Mitaka nova-net DevStack without cells.

Change-Id: I74f67ac8eb12c7a649ddcbd7979cf745fb35cc0c