Updated openstack/openstack Project: openstack-infra/project-config  9c373f798ef3c8849d48eab7041cd9d91c722485
Clean up stable jobs due to driverfixes filters

Some driverfixes branch filters were added later in the
yaml file causing previously filtered jobs to get re-added
to the list of jobs to run. This moves some of the branch
filtering to the earlier filter in the file to get rid of
the later attempts.

Change-Id: Id9f40c665a912afde8ec8051ab3ea5192a51bde9