Updated openstack/openstack Project: openstack-infra/irc-meetings  e786a2292927aa1e3f2510d49f959465fa235bed
Remove the keystone/horizon meeting

We used this meeting heavily in the Ocata development cycle to smooth
out several issues between keystone and horizon. It was determined by
both teams that a weekly cadence is no longer necessary and we plan
to call meetings on an as-needed basis. The full thread can be viewed
here:

  http://lists.openstack.org/pipermail/openstack-dev/2017-April/115761.html

Change-Id: Ic51b7b475ec61cd482ebf6d92d712f1c52a0c054