Updated openstack/openstack Project: openstack-infra/tripleo-ci  bb997040e1c3bba24cae1e4edcf421ebe3095407
Fix path for the multinode job when running the major upgrade.

This submission fixes the environment path when running the
major upgrade.

We have two multinodes jobs, one running a major upgrade and
other not doing it.

Currently we make no distinction building the MULTINODE_ENV_PATH
depending if an upgrade will run.

This submission moves the conditional already used for the upgrades
scenarios to build the MULTINODE_ENV_PATH also for the multinode upgrades
job.

The current deployment builds the paths like:
-e /usr/share/openstack-tripleo-heat-templates/ci/environments/multinode.yaml
--templates /opt/stack/new/ocata//usr/share/openstack-tripleo-heat-templates

And after applying it:
-e /opt/stack/new/ocata/usr/share/openstack-tripleo-heat-templates/ci/environments/multinode.yaml
--templates /opt/stack/new/ocata//usr/share/openstack-tripleo-heat-templates

Closes-Bug:1687908
Change-Id: If88f403c85b79bd896a24c7816486709bd67706f