Updated openstack/openstack Project: openstack/neutron  fde6710515681a1cbe2a2af33601f01d48a832fd
use MechanismDriver from neutron-lib + shim

The ml2 MechanismDriver is now in neutron-lib along with its associated
constants. This patch switches over to the lib versions of those, but
leaves a shim of the MechanismDriver that just ref's the driver from
lib. This shim allows our broad consumer base of the driver to switch
over at their leisure.

NeutronLibImpact

Change-Id: I99e3de6d933a1bb341394f85415fb07306a82a01