Updated openstack/openstack Project: openstack/requirements  5d2a06e90e2476a9a53392b22815ff266eaf76a7
bump neutron-lib to 1.6.0

The upper constraints for neutron-lib have already been bumped for 1.6.0
in Ic5f6099a871a58bbc571fb40de48f7d3178fce7d, but we'd like to see the
requirement change propagated into all consumers so that we can begin
using the changes in neutron-lib between 1.4.0 and 1.6.0. This patch
bumps neutron-lib to 1.6.0 so consuming requirements.txt will be updated
allowing us to use the latest release in neutron and other networking
projects.

Change-Id: I16c22053456d02a83f27b42d6649c1180efedade