Updated openstack/openstack Project: openstack/tempest  892088b0cd4e805e0f923e14b03b0be6eccf1cc4
Add missing IPs validation to test_rebuild_server

The test_rebuild_server test verifies several properties in a server
that has been rebuilt, however the IP addresses are never checked,
and after a rebuild the IP addresses should not change, so they should
be validated also by this test. This commit adds that validation to the
test.

Change-Id: If4baf51f1d36fb2298a01c28f6d55f851533cac4
Closes-Bug: 1600372