Updated openstack/openstack Project: openstack-infra/project-config  9f1507eabc4552b370963042ec6c5f2f2ac38176
sahara: refactor image building, new builders

Generalize the existing image building jobs and distinguish
the repository (sahara-image-elements for the existing
build system, sahara for the new build/validation method).

Right now only two new jobs are defined for sahara (based on
sahara-image-pack). This matches the current status of the
almost ready (under review) sahara-image-pack builders.
As the code for the builders are not ready, the new jobs are
only defined in the experimental queue.

The script called by the job has the same name in both
repositories (sahara-image-elements and sahara).

The refactor also consistently changes the name of all non-voting
jobs by adding the -nv suffix (which makes them non-voting by
default with some magic) and removes two non-voting jobs
which should not have been in the gate queue.

Change-Id: I5fc00c4c8f97ecaa9e48e2e76a64ddc6d68d1720