Updated openstack/openstack Project: openstack/nova  fb5b1d0c90fefa263c4abfa8287133a91ce06d70
placement: get providers sharing capacity

This is the first phase of adding support for shared resource providers to the
scheduler. We add a new function "_get_providers_with_shared_capacity()" to
nova/objects/resource_provider.py that accepts a resource class identifier and
a requested amount of that resource class and we return a list of identifiers
for resource providers that a) are marked as sharing their inventory via an
aggregate (with the new MISC_SHARES_VIA_AGGREGATE trait) and b) have enough
capacity to service the requested amount of the resource.

In followup patches, the ResourceProviderList.get_all_by_filters() method will
call _get_providers_with_shared_capacity() to identify those resource providers
that it can consider as providing a resource for a request when a compute node
provider doesn't have inventory of that resource.

Only included is a functional test, since this is all SQL-fu and there's no
good way to unit-test any of that.

Change-Id: Ie8802c524b96998e10d436898ca78ef30567fef8
blueprint: shared-resources-pike