Updated openstack/openstack Project: openstack/manila  490ddb7c1b5c43a335a3e084d09c5822f55e327d
Change to share access list API

When we do manila access-list, the 'created_at' and 'updated_at' time
should be shown to the user. And then the user could determine which
is the recent access rule. As the design, the recent access-allow
rule caused the error access-status of share instance.

APIImpact
Closes-Bug: #1682795
Change-Id: Iad6070d60ec77b7de9cc9679cfa7478876084da1