Updated openstack/openstack Project: openstack/openstack-manuals  bbcbb97b1c7bed38e7cd69097122e62e9c4dc27d
Add link to training-labs to the install-guide

Since Icehouse, the training-labs scripts have offered an automated
method of installing the exact cluster described in the install-guide
for every OpenStack release. In order to make this option easier to
find, this changeset adds a brief note and a link from the install-guide
to training-labs.

Change-Id: Iada7ad7b0d52176d1a1ad405769882b24a6d1f15