Updated openstack/openstack Project: openstack-infra/storyboard-webclient  ccb66494e93205f262214b64bb6801fcde06cadc
Handle expired refresh tokens by reloading the page

If a user's refresh token expires when their access token is also
expired, they cannot obtain a new access token without authenticating
with the openid provider.

Currently StoryBoard doesn't handle this situation, and the session
ends up in a state whereby the access token is invalid and most
requests cause errors due to this.

This commit hackily fixes this by reloading the page if the request
for a new access token fails, such as when the refresh token is
expired.

Change-Id: Ied597133a76e5e61677fafd7bfb35d4107e0836c