Updated openstack/openstack Project: openstack/oslo.config  47d7c1c090fdbc0dabd95b043a82eabdc201dead
Handle multi-line default values in olso-config-generator

The oslo.config file format permits multi-line values for options,
however all the lines except for the first one need to be indented.

Right now oslo-config-generator fails to take that into account, and
instead outputs the formatted value as is, adding a comment character
only to the first line, and not commenting out or indenting the
subsequent lines.

This patch tries to fix that by adding a comment character and some
indentation to every newline in the formatted value.

Change-Id: I33381c8cfb59901ce6dfc07bd662f2edb56531dd
Closes-bug: #1689594