Updated openstack/openstack Project: openstack/horizon  978e67f1447f34dcb848aa0c6f220d01ffb92993
Add a new management cmd to pull translations

This patch adds a new command to pull translation catalogs directly from
Zanata master. This is so that translaters can easily see their changes.

Its general usage will be as part of an infra & OSA tool, but it makes
sense to keep the logic in Horizon for those familiar with Horizon and
i18n.

Change-Id: I19e493dc1debb48b68a56b38c533429d63600a56