Updated openstack/openstack Project: openstack-infra/tripleo-ci  29e90cc9c0031439ce9ad9da8bc9f3f1e00e8d42
Support running toci_gate_test-oooq.sh locally

These are various improvements to tripleo-ci (squashed into a single
commit to save CI resources) that enable running the tripleo-ci entry
point script which is now toci_gate_test-oooq.sh locally in a standalone
environment of vm's not spawned by nodepool.

This is useful to developers trying to reproduce CI locally for
development and testing.

The changes are:
- Default $REMAINING_TIME in common_vars.bash
- Don't hardcode the jenkins user everywhere
- Improved initial package handling:
    Before blindly yum installing the packages needed by these scripts in a
    few spots, check if the packages are already installed with just the rpm
    command. This can end up saving a decent amount of time when iterating
    on running these scripts locally. Also installs wget.
- Reinstall python-requests
    When locally running toci_gate_test.sh, python-requests needs to be
    reinstalled since toci_gate_test.sh removes the python module as it
    was pip installed on the infra provided image. Also updates the
    logic to handle the python-requests -> python2-requests transition.

Change-Id: I48cd1e5f53b86455992b83248d57f5338c29a7ca