Updated openstack/openstack Project: openstack/ceilometermiddleware  e2bf485044d8f3743da9298a9e461c5808be31f3
retrieve project id to ignore from keystone

We currently allows only project uuid, but this is a pain for deployer.
Also the default is a project name which doesn't work...

This change queries keystone to retrieve project ids when the
ignore_projects list are names.

By default, if auth_type is not set, we keep the previous behavior to
not query keystone.

Change-Id: I270d080d3e65eb6b0cd823498a4dd37389c49221