Updated openstack/openstack Project: openstack-infra/shade  b0e2e5568953349c7085cfc07f02060c4aa3d716
Move mocks of designate API discovery calls to base test class

As API discovery call has to be mocked in each Designate's
test case it's better to move it to base test class and reuse
later.
This commit moves it to tests.unit.base.RequestsMockTestCase that
it can be added to mocks in setUp() method before each test.

Change-Id: I3f00c2d51619d634e4e44f3ebf276949544c8575